Change typing

+++ JumpRoyale/src/Jumper.cs
@@ -19,7 +19,7 @@ public class Jumper(PlayerData playerData)
     /// <summary>
     /// Realtime updated Y-position of this jumper - controlled by JumperScene.
     /// </summary>
-    public float CurrentHeight { get; set; }
+    public double CurrentHeight { get; set; }
 
     public void Initialize()
     {
